Display requested times in request list
This, along with commits 7a9b0ac and 9c420e7 should at least partially resolve
issue #172 as I undersand it. I think the problem is that the requested dates
weren't being displayed, even though they were stored in the database.

There is another side to the story: The due_date may still report a premature
overdue status, because it's set relative to the date_created. The set_due_date
method is going to have to become smarter to take into account the time of day
that the request was made. For now, displaying the time at least allows users
to spot check whether the request is ACTUALLY overdue.